Se cambian los mensajes de salida y se quita el console.log

+++ src/controllers/ProjectController.ts
@@ -7,7 +7,10 @@ export class ProjectController {
     static createProjects = async (req: Request, res: Response) => {
 
         const project = new Project(req.body);
-        console.log(project)
+
+        // Asignar el usuario que crea el proyecto
+        project.manager = req.user.id;
+
         try {
             await project.save();
             // await Project.create(req.body);  Otra forma de como guardar un proyecto
@@ -20,7 +23,13 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({});
+            const projects = await Project.find({
+                // ! Validar el rol del usuario
+                // $or: [
+                //     { manager: { $in: req.user._id } },
+                //     { team: { $in: req.user._id } } 
+                // ]
+            });
             res.json(projects);
         } catch (error) {
             console.log(error)
@@ -28,17 +37,22 @@ export class ProjectController {
     }
 
     static getProjectById = async (req: Request, res: Response): Promise<void> => {
+        const { id } = req.params
         try {
-            const project = await Project.findById(req.params.id).populate("tasks");
+            const project = await Project.findById(id).populate('tasks')
             if (!project) {
-                const error = new Error("Proyecto no encontrado");
-                res.status(404).json({ msg: error });
+                const error = new Error('Proyecto no encontrado')
+                res.status(404).json({ error: error.message })
                 return;
             }
-            res.json(project);
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no válida')
+                res.status(404).json({ error: error.message })
+                return
+            }
+            res.json(project)
         } catch (error) {
-            console.log(error);
-            res.status(500).send("Hubo un error en el servidor");
+            console.log(error)
         }
     }
 
@@ -52,6 +66,12 @@ export class ProjectController {
                 res.status(404).json({ error: error.message });
                 return;
             }
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no válida - Manager no válido')
+                res.status(404).json({ error: error.message })
+                return
+            }
+
             project.clientName = req.body.clientName;
             project.projectName = req.body.projectName;
             project.description = req.body.description;
@@ -67,12 +87,19 @@ export class ProjectController {
         const { id } = req.params;
         try {
             const project = await Project.findById(id);
-            await project.deleteOne();
             if (!project) {
                 const error = new Error("Proyecto no encontrado");
                 res.status(404).json({ msg: error });
                 return;
             }
+            
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no válida - Manager no válido')
+                res.status(404).json({ error: error.message })
+                return
+            }
+
+            await project.deleteOne();
             res.send("Proyecto eliminado correctamente");
         } catch (error) {
             console.log(error);
